+++ tests/utils.ts
@@ -1,4 +1,3 @@
-import { assertSnapshot } from "../deps/snapshot.ts";
 import lume from "../mod.ts";
 import Server from "../core/server.ts";
 import { basename, fromFileUrl, join } from "../deps/path.ts";
@@ -100,8 +99,7 @@ export async function assertSiteSnapshot(
   const { pages, files } = site;
 
   // To-do: test site configuration
-  await assertSnapshot(
-    context,
+  await context.assertSnapshot(
     {
       formats: Array.from(site.formats.entries.values()).map((format) => {
         // deno-lint-ignore no-explicit-any
@@ -183,8 +181,8 @@ export async function assertSiteSnapshot(
   });
 
   // Test static files
-  await assertSnapshot(context, normalizedFiles);
-  await assertSnapshot(context, normalizedPages);
+  await context.assertSnapshot(normalizedFiles);
+  await context.assertSnapshot(normalizedPages);
 }
 
 export function getServer(
@@ -221,7 +219,7 @@ export async function assertResponseSnapshot(
   const body = await response.text();
   const headers = Object.fromEntries(response.headers.entries());
 
-  await assertSnapshot(context, {
+  await context.assertSnapshot({
     request: request.url,
     method: request.method,
     status: response.status,