+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - `favicon` plugin: Don't override favicons added explicitly with `site.add()` or `site.copy()`.
 - Archetypes are loaded from the cwd, not the `src` folder.
 - Moved `Logger` class to the `core/` folder.
+- Removed `@std/testing/snapshot` dependency in favor of buil-in `t.assertSnapshot()` function introduced in Deno 2.9
 
 ### Fixed
 - `icons` plugin: Added missing type for `spriteIcon` helper [#852].